Minor cleanups and fixes
